Change encryption preference's summary accordingly
With this change a summary to the encryption preference was added.

https://android.googlesource.com/platform/packages/apps/Settings/+/415ce9a77d44732966fbab641d1ecaa8e8688381

The problem is that even if you're not encrypted, it says that your phone is.
Is very misleading to the user that don't bother to enter the actual preference
and want to see if they are. This fixes that.

Test: Go into Settings/Security & Location

Change-Id: Iaeaffd340f2e248748f244a947aa71eb8ad4ab94
Signed-off-by: Eliminater74 <eliminater74@gmail.com>